splitting off code to seperate library
